+++ examples/diagram_generators/__init__.py
@@ -1,0 +1,15 @@
+"""
+Shared diagram generation utilities for creating examples and showcases.
+"""
+
+from .graphviz_examples import get_graphviz_examples
+from .mermaid_examples import get_mermaid_examples
+from .plantuml_examples import get_plantuml_examples
+from .showcase_generator import generate_unified_showcase
+
+__all__ = [
+    "generate_unified_showcase",
+    "get_mermaid_examples",
+    "get_plantuml_examples",
+    "get_graphviz_examples",
+]
